+++ Opt/Some.cs
@@ -7,30 +7,29 @@ namespace ComradeVanti.CSharpTools
     ///     An optional value that is present
     /// </summary>
     /// <typeparam name="TValue">The type of the contained value</typeparam>
-    internal sealed class Some<TValue> : IOpt<TValue>
+    internal sealed class Some<TValue> : ISome<TValue>
     {
-
-        private readonly TValue value;
+        public TValue Value { get; }
 
 
         public Some(TValue value) =>
-            this.value = value;
+            Value = value;
 
 
         public TResult Match<TResult>(Func<TValue, TResult> onSome, Func<TResult> _) =>
-            onSome(value);
+            onSome(Value);
 
 
         public override bool Equals(object obj) =>
             obj is IOpt<TValue> other
-            && other.Match(otherValue => Equals(value, otherValue),
+            && other.Match(otherValue => Equals(Value, otherValue),
                            () => false);
 
         public override int GetHashCode() =>
-            value.GetHashCode();
+            Value.GetHashCode();
 
         public override string ToString() =>
-            $"Opt {{ {value} }}";
+            $"Opt {{ {Value} }}";
 
     }