+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ﻿# Changelog
 
+## Upcoming
+
+### Added
+
+- ISome and INone interfaces in order to support pattern matching on IOpt instances. 
+  Also added relevant documentation.
+
 ## 2.2.0
 
 ### Added